[IMP] pos_loyalty: noisy active programs section
Before this commit
==================
The sections about Discount & Loyalty shows a lot of redundant or irrelevant
information for the user and it leaves no place for order lines.

After this commit
=================
Just keep one section named "Discounts" with the description of each
"reward" applied. To leave more place for order lines, "Discount" section
is set beside of loyalty points section.

task - 2980232
