+++ addons/pos_loyalty/static/src/xml/OrderSummary.xml
@@ -2,33 +2,48 @@
 <templates id="template" xml:space="preserve">
     <t t-name="OrderSummary" t-inherit="point_of_sale.OrderSummary" t-inherit-mode="extension" owl="1">
         <xpath expr="//div[hasclass('summary')]" position="after">
-            <div class="summary clearfix">
-                <t t-set="_loyaltyStats" t-value="getLoyaltyPoints()"/>
-                <t t-foreach="_loyaltyStats" t-as="_loyaltyStat" t-key="_loyaltyStat.couponId">
-                    <t t-if="_loyaltyStat.points.won || _loyaltyStat.points.spent">
-                        <div class='loyalty-points'>
-                            <div class='loyalty-points-title'>
-                                <t t-esc="_loyaltyStat.points.name"/>
+            <div>
+                <div class="summary">
+                    <t t-set="_loyaltyStats" t-value="getLoyaltyPoints()"/>
+                    <t t-foreach="_loyaltyStats" t-as="_loyaltyStat" t-key="_loyaltyStat.couponId">
+                        <t t-if="_loyaltyStat.points.won || _loyaltyStat.points.spent">
+                            <div class='loyalty-points'>
+                                <div class='loyalty-points-title'>
+                                    <t t-esc="_loyaltyStat.points.name"/>
+                                </div>
+                                <t t-if='_loyaltyStat.points.won'>
+                                    <div class="loyalty-points-won">
+                                        <span class='value'>+<t t-esc='_loyaltyStat.points.won'/></span>
+                                    </div>
+                                </t>
+                                <t t-if='_loyaltyStat.points.spent'>
+                                    <div class="loyalty-points-spent">
+                                        <span class='value'>-<t t-esc='_loyaltyStat.points.spent'/></span>
+                                    </div>
+                                </t>
+                                <div class='loyalty-points-total'>
+                                    <span class='value'><t t-esc='_loyaltyStat.points.total'/></span>
+                                </div>
                             </div>
-                            <t t-if='_loyaltyStat.points.won'>
-                                <div class="loyalty-points-won">
-                                    <span class='value'>+<t t-esc='_loyaltyStat.points.won'/></span>
+                        </t>
+                        <t t-else="">
+                            <div></div>
+                        </t>
+                    </t>
+                </div>
+                <div>
+                    <t t-set="loyaltyData" t-value="getActiveRewards()"/>
+                    <div class="active-programs">
+                        <div t-if="loyaltyData.activeRewards.length">
+                            <div class="title">Discounts</div>
+                            <t t-foreach="loyaltyData.activeRewards" t-as="reward" t-key="reward.id">
+                                <div class="active-reward">
+                                    <t t-esc="reward.name"/> <t t-if="reward.code">(<t t-esc="reward.code"/>)</t>
                                 </div>
                             </t>
-                            <t t-if='_loyaltyStat.points.spent'>
-                                <div class="loyalty-points-spent">
-                                    <span class='value'>-<t t-esc='_loyaltyStat.points.spent'/></span>
-                                </div>
-                            </t>
-                            <div class='loyalty-points-total'>
-                                <span class='value'><t t-esc='_loyaltyStat.points.total'/></span>
-                            </div>
                         </div>
-                    </t>
-                    <t t-else="">
-                        <div></div>
-                    </t>
-                </t>
+                    </div>
+                </div>
             </div>
         </xpath>
     </t>
